+++ ion/src/blackbox/ion.cpp
@@ -1,8 +1,8 @@
 #include <stdint.h>
 #include <ion.h>
 
-void Ion::msleep(long ms) {
+void Ion::msleep(uint32_t ms) {
 }
 
-void Ion::usleep(long us) {
+void Ion::usleep(uint32_t us) {
 }
